配合hcom_topo_info的修改修改ut打桩



# message auto-generated for no-merge-commit merge:
!2911 merge fix_hcom_topo_info_stub into master

配合hcom_topo_info的修改修改ut打桩

Created-by: xutianze
Commit-by: xutianze
Merged-by: turing_project1
Description: ## 描述
<!--在这里详细描述你的改动，包括改动的原因和所采取的方法。-->

## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在这里提供Issue链接。例如：关联Issue #123-->
<!-- 如果这个PR是为了解决特定的问题单，请在这里描述问题单单号。-->

## 测试
<!--描述进行了哪些测试来验证你的改动。包括但不限于二级冒烟、算子泛化等。-->

## 文档更新
<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->

## 类型标签
<!--  [x] 表示选中 -->
- [ ] Bug修复
- [ ] 新特性
- [ ] 性能优化
- [ ] 文档更新
- [ ] 其他，请描述：




See merge request: cann/ops-transformer-dev!2911

+++ tests/ut/framework_normal/common/mc2_hcom_topology_mocker.h
@@ -18,19 +18,19 @@
 #include <unordered_map>
 
 namespace Mc2Hcom {
-using MockValues = std::vector<std::pair<const char*, uint32_t>>;
+using MockValues = std::vector<std::pair<const char*, uint64_t>>;
 
 class MC2HcomTopologyMocker {
 public:
     static MC2HcomTopologyMocker& GetInstance();
-    void SetValue(const char* key, uint32_t value);
+    void SetValue(const char* key, uint64_t value);
     void SetValues(const MockValues& values);
-    uint32_t GetValue(const char* key, uint32_t defaultValue) const;
+    uint64_t GetValue(const char* key, uint64_t defaultValue) const;
     void Reset();
 
 private:
     MC2HcomTopologyMocker() = default;
-    std::unordered_map<std::string, uint32_t> mockValue_;
+    std::unordered_map<std::string, uint64_t> mockValue_;
 };
 
 }
